requirements.txt
fastapi
uvicorn
tensorflow
numpy
python-multipart
pillow
gradio